Cirrus: Rearrange tasks + upload binary artifacts
As a minor, readability improvement, rearrange tasks in sequence order.
Also ensure the "validate" task is a dependency of "success".

To help support downstream testing with a rapidly evolving aardvark and
netavark, make the compiled and tested binary available for download.
In the future this should be replaced by use of packaged binaries,
pre-installed into the test VM.

Signed-off-by: Chris Evich <cevich@redhat.com>