dns, server: don't trigger kill_switch inside receiver block
Why ?
It seems there is a race if interface is removed from below and closing
sockets won't stop the event loop. So lets de-couple thread killing and
networking logic.

Solution:
Reciver is already async so leverage that and move kill_switch to an
event loop which is not dependent on receiver.

Following solution makes sure that our thread management is not
dependent on any networking code. If we want to kill the thread we can
kill it.

Signed-off-by: Aditya R <arajan@redhat.com>