server: use async_broadcast channel instead of delay in event loop
We are using all most all of CPU with the approach where we are trying
to read value of kill switch in a delayed loop. This approach looks bad.

Instead start listening to async broadcast channel as a kill switch
mechanism, this will make sure that we dont eat CPU for no reason.

Signed-off-by: Aditya R <arajan@redhat.com>