Add "IBM" use case 1/2
Verify that one of the IBM use cases where we have three networks (A, B,
C).  And the a container is created on A and one on C, referred to as
A1 and B1.  Then A1 and C1 do a network connect and join network B and
they each add and alias.

Signed-off-by: Brent Baude <bbaude@redhat.com>