tests: fix slirp flake
When we create a new netns with unshare we have to wait until it
completes before using the ns. When we run things in the background with
`&` we always need to check that the process is ready before using it.
We now have ready checks for unshare and slirp4netns so it should no
longer flake even on slow systems.

Also only setup slirp4netns when needed for the test to safe some setup
time.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>