config: use multiple ips per line
We have to support several ipv4 and v6 addresses per container. The
current way of creating the configs in aardvark are just broken IMO. We
should never have to duplicate the container id and aliases just to set
more ips.

This patch is fully backwards compatible and will allow much better
config generation in netavark.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>