aardvark main: change error reporting
Never report an error when the syslog init fails, there are systems
without syslog (e.g. inside containers) and we should not spam a random
expected messages in this context.

Also add aardvark-dns to the error context as this is printed to stderr
which is inherited from the podman command so it is not clear which
process wrote it otherwise.

Fixes https://github.com/containers/podman/issues/19809

Signed-off-by: Paul Holzinger <pholzing@redhat.com>