fix duplicated IP CI flake
The test setup was wrong, choosing a ip randomly had the result that
sometimes two containers ended up with the same ip which resulted in
aardvark-dns returning a wrong response.

To fix this lets make sure to not assing the same ip more than once by
using a counter.

Fixes #411

Signed-off-by: Paul Holzinger <pholzing@redhat.com>