Internal networks cannot make external DNS requests
Internal networks cannot connect to the internet thanks to
routing, but they can connect to Aardvark, which will happily
forward their DNS requests to the internet.

This could theoretically be used to build a data-exfiltration
sidechannel.

Fix this by identifying internal networks with a filename suffix
(using a character disallowed in actual network names to ensure
we don't conflict with another network) and explicitly setting
their DNS servers to an empty list (and refusing to set
per-container DNS at all). We could actually error on finding DNS
servers in an internal network, but silently ignoring prevents
possible compatibility issues with Netavark.

Signed-off-by: Matt Heon <mheon@redhat.com>