fix reverse ipv6 lookup test flake
The test assumed that we have a fdxx:... addresses which is what we
should be getting normally, however the ipv6 subnet generator had a bug
that caused it to create fdx:... sometimes which is different as ipv6
allows us to omit leading zeros, thus we actually have 0fdx which is
not in the range we want to use at all.

So simply fix this by creating the proper subnet so that if the random
number was less than 16 we add the leading zero add the proper place
after fd.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>