[skip-ci] Packit: separate `packages` key for rhel jobs
This should prevent status reporting issues as encountered in
https://github.com/containers/aardvark-dns/pull/458 . The possible cause
for the error was packit getting confused between two separate copr jobs
with the same parameters but different targets.

Having a separate `packages` key for rhel jobs should make it easier for
packit to do the right thing.

This should also make it possible for CentOS integration tests to run
automatically on all PRs and not need manual triggering by a maintainer.
Those were not triggered automatically on
https://github.com/containers/aardvark-dns/pull/456 while the Fedora
integration tests were.

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>