coredns: move upstream resolvers detection later
We only care about upstream resolvers if we actually have to forward a
request. As such move this down this safes some map lookups in the best
case. Also fix the use resolv_conf::Config wich is not needed at all and
just confusing as we only need the nameservers. And rename the var to
clear state what they are upstream resolvers.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>