server: improve parent <-> child error handling
The code currently doesn't check for most child errors as it basicaly
immediately wrote the ready byte. Also it never considered the fact that
the child can exit early and in that case it would have blocked forever
because the parent never closed the writing side before call read().

This fixes that part, however it still reports success way to early. It
should wait for the first bind but that needs more complicated changes.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>