replace signal-hook with tokio::signal
We already have tokio as dependency anyway so we might as well just use
it for the signal code as well.
They both internally depend on signal-hook-registry so it isn't a big
gain in terms of used code but it seems cleaner to stick to the tokio
API where possible and the resulting binary is about 14 kb smaller.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>