serve: read nameservers once
When we span multiple listeners at once we each one read resolv.conf.
This of course makes no sense as reading once is more than enough.

this fixes that, however there is still the problem that these server
are never read again after the per network server is started which means
if resolv.conf changes aardvak-dns does not update it. But the old code
had the same issue so I ignore it for now and come back to that later.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>