coredns: drop pointless name var
It was only used in one log statement and not used elsehwere. And
logging a truncated name doesn't make sense either just log the full
network name which is what users would care about anyway.

This removes the need to the parsing code of the name allowing us to
drop the Result and remove some error handling.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>