server: split out main loop into new function
The goal here is to make sure the first startup can happen outside the
loop so we can handle the error and exit to return a useful error on the
first bind/config parsing to netavark which can then return it to podman.
The bind problem doesn't work right now but this will be fixed in the
next commit.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>