return bind error to caller on first start
When we fail to bind the first time we are started we must report this
back to the user so they know why dns isn't working.
In order to capture all errors we need to make use of the multi error
type so we can list all bind failures not just the first one.
Also change the port to u16 because a port cannot be higher than that it
safes us from making a failable conversion later on in the stack.

With that and using the patched netavark for the error handling podman
shows now this
```
Error: netavark: IO error: Error while applying dns entries: IO error: aardvark-dns failed to start: Error from child process
Error starting server failed to bind udp listener on 10.89.1.1:53: IO error: Address already in use (os error 98)
failed to bind udp listener on 10.89.0.1:53: IO error: Address already in use (os error 98)
```

The error message is not exactly nice working so we should fix some of
that later but at least we not can show the bind errors correctly.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>