tcp: add timeout to connection
Add a 3s timeout to all tcp connections, we do not want clients to keep
the connections open forever. Also do not allow more than one message
per connection.

The API is a bit weird, we first get the message then have to poll again
where it return None otherwise the reply will not be send.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>