backend: return simple Vector in lookup()
There is no use of the enum, the caller just checks if it gets ips or
not so we can make this much simpler.

Now there are a lot of unit test using this so I had to fix them. While
fixing I decided it woul dbe a godd idea to also change the format they
are written as this all seems super complicated for no reason there.
We can direclty check the vector in assert_eq!() and do not need the
complicated destructuring. And just use expect() directly over matching
error just to panic anyway. Also parsing the same ips over and over
doesn't look nice either.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>