dns: limit to 3 resolvers and use better timeout for them
The default timeout was 5s and as 5s is the default for most clients we
may try more upstream servers but at that point the client already
considered this a timeout and no longer accepts answers. As such if we
want to try more reolvers in case the first one is offline/not working
then we need a lower timeout. Let's only try 3 as glibc does and then
just divide 5s through the number of resolvers, i.e. 1, 2, 3. This
should make the fall back work.

Fixes #482

Signed-off-by: Paul Holzinger <pholzing@redhat.com>