serve: parse resolv.conf ourselves
The resolv.conf parsing lib is super strict and does not allow unknown
options, given we do not care about options search domains or really
anything else in this file parse it ourselves with very lax rules.

Basically we try to find the nameserver lines. Anything else is ignored,
the only error we produce is if we fail to parse the nameserver ip
address.

Fixes #418
Fixes https://issues.redhat.com/browse/RHEL-57695

Signed-off-by: Paul Holzinger <pholzing@redhat.com>