support ipv6 link local addresses in resolv.conf
The recent change to move parsing resolv.conf into our code caused a
regression with link local addresses. We did not split "%" correctly due
a typo. While the fix for that is simple ignoring the zone part as we
did before will not work as the zone must be given for link local
addresses.

And as it turns out it does not need much changes to actually support
the zone part properly. We can just pass the SocketAddr type around. The
only problem there is that we must know the numeric id and not the
interface name in the struct. To resolve name -> id we can simply call
if_nametoindex that is provided by libc which is safely wrapped in the
nix crate.

Fixes 22293ef ("serve: parse resolv.conf ourselves")
Fixes #535

Signed-off-by: Paul Holzinger <pholzing@redhat.com>