test: use ncat not nc
nc can be provided by either ncat (nmap) or netcat (OpenBSD), we only
work with the nmap version so make sure we always use that one and not
the short alias which can be resolved to either one.

It is not clear to me what changed on rawhide but it seems on the tmt
env nc is preferred even though we have nmap-ncat installed.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>