deps: update nix to 0.30.1
The dup2 API has changed and now requires an OwnedFd as traget fd, that
is not what we have or want. An OwnedFd is closed on drop so even if I
would create one it would be dropped at the end of the function which
would then close stdin/out/err which would be clearly wrong.

To accommodate the common use case of replacing the std stream nix added
the dup2_stdin/stdout/stderr functions instead which we now use.

Also read no longer accepts a raw fd either.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>