test: replace ncat with socat
ncat keeps causing issues with different versions having different
behaviors. And many distro ship different incompatible versions so it is
hard to make the test work everywhere consistently.

To address replace ncat with socat. Socat seems to much more stable and
we have been using it in podman for pasta tests for a while without
problems.

Fixes: #622

Signed-off-by: Paul Holzinger <pholzing@redhat.com>